Update Vires in Numeris.py

+++ Vires in Numeris.py
@@ -17,21 +17,20 @@ log = logging.getLogger(__name__)
 class ViN(IStrategy):
     INTERFACE_VERSION = 2
 
+    write_to_txt = False
     f_buys = './user_data/vinbuys.txt'
     f_trades = './user_data/vintrades.txt'
     write_to_csv = False
     df_csv = './user_data/df.csv'
     buy_time_periods = (13, 14, 15, 16, 17)
-    indicator_range = range(3, max(buy_time_periods)+1)
-    has_bt_agefilter = True
-    has_downtime_protection = False
-    min_vol_candle: int = 1000
-    min_vol_1h: int = 10000
-    max_concurrent_buy_signals: int = 15
+    indicator_range = range(2, max(buy_time_periods)+1)
+    min_vol_candle: int = 1500
+    min_vol_6_candles: int = min_vol_candle * 4
+    # max_concurrent_buy_signals: int = 15
     custom_buy_info = {}
 
     minimal_roi = {"0": 10}
-    stoploss = -0.99 #-0.08
+    stoploss = -0.08
     stoploss_on_exchange = False # True
     trailing_stop = False
     use_custom_stoploss = False
@@ -40,22 +39,16 @@ class ViN(IStrategy):
     use_sell_signal = True
     sell_profit_only = False
     ignore_roi_if_buy_signal = True
-    startup_candle_count: int = 144
+    startup_candle_count: int = 36
 
     def populate_indicators(self, df: DataFrame, metadata: dict) -> DataFrame:
         if len(df) < self.startup_candle_count:
             return df
 
-        # positive = (df['close'] - df['open']).ge(0)
-        # bodysize = (df['close'] / df['open']).where(positive, 0 - df['open'] / df['close'])
-        # df['uppertail'] = (df['high'] / df['close']).where(positive, df['high'] / df['open'])
-        # df['lowertail'] = (df['open'] / df['low']).where(positive, df['close'] / df['low'])
-
         df_closechange = df['close'] - df['close'].shift(1)
         for i in (1, 2, 3):
             df['updown'] = np.where(df_closechange.rolling(window=i, min_periods=1).sum().gt(0), 1, np.where(df_closechange.rolling(window=i, min_periods=1).sum().lt(0), -1, 0))
             df[f"streak_{i}"] = df['updown'].groupby((df['updown'].ne(df['updown'].shift(1))).cumsum()).cumsum()
-
         df['streak_min'] = df[['streak_1', 'streak_2', 'streak_3']].min(axis=1)
         df['streak_max'] = df[['streak_1', 'streak_2', 'streak_3']].max(axis=1)
         df.drop(columns=['updown', 'streak_1', 'streak_2', 'streak_3'])
@@ -64,17 +57,17 @@ class ViN(IStrategy):
         for i in self.indicator_range:
             df[f"mom_{i}"] = ta.MOM(df, timeperiod=i)
             upp, mid, df[f"mom_{i}_low"] = ta.BBANDS(df[f"mom_{i}"], timeperiod=i, nbdevup=2.0, nbdevdn=2.0, matype=0)
-            df[f"rsi_{i}"] = ta.RSI(df, timeperiod=i)
             df[f"mfi_{i}"] = ta.MFI(df, timeperiod=i)
             df[f"cti_{i}"] = cf['index'].rolling(i).corr(cf['close'], method='spearman')
 
-        df['volume_12'] = df['volume'].rolling(12).sum()
-        if self.config['runmode'].value in ('live', 'dry_run'):
-            if self.has_downtime_protection:
-                df['live_data_ok'] = (df['volume'].rolling(window=72, min_periods=72).min() > 0)
-        else:
-            if self.has_bt_agefilter:
-                df['bt_agefilter'] = df['volume'].rolling(window=self.startup_candle_count, min_periods=self.startup_candle_count).count()
+        positive = (df['close'] - df['open']).ge(0)
+        # bodysize = (df['close'] / df['open']).where(positive, 0 - df['open'] / df['close'])
+        # df['uppertail'] = (df['high'] / df['close']).where(positive, df['high'] / df['open'])
+        df['lowertail'] = (df['open'] / df['low']).where(positive, df['close'] / df['low'])
+        df['closechange'] = df['close'].pct_change()
+
+        df['volume_6'] = df['volume'].rolling(6).sum()
+        df['candle_count'] = df['volume'].rolling(window=self.startup_candle_count, min_periods=self.startup_candle_count).count()
 
         return df.copy()
 
@@ -84,14 +77,9 @@ class ViN(IStrategy):
 
         df.loc[:, 'buy_tag'] = ''
         conditions = []
+        conditions.append(df['candle_count'].ge(self.startup_candle_count))
         conditions.append(df['volume'].ge(self.min_vol_candle))
-        conditions.append(df['volume_12'].ge(self.min_vol_1h))
-        if self.config['runmode'].value in ('live', 'dry_run'):
-            if self.has_downtime_protection:
-                conditions.append(df['live_data_ok'])
-        else:
-            if self.has_bt_agefilter:
-                conditions.append(df['bt_agefilter'].ge(self.startup_candle_count))
+        conditions.append(df['volume_6'].ge(self.min_vol_6_candles))
 
         for i in self.buy_time_periods:
             buy_condition = []
@@ -101,11 +89,11 @@ class ViN(IStrategy):
                 buy_condition.append(df['streak_min'].le(-i))
             else:
                 buy_condition.append(df['streak_min'].eq(-i))
-            buy_condition.append((df[f"mom_{i}"] / df[f"mom_{i}_low"]).between(1.05, 1.25))
-            buy_condition.append(df[f"rsi_{i}"].between(10, 10 + i))
+            buy_condition.append((df[f"mom_{i}"] / df[f"mom_{i}_low"]).between(1.1, 1.2))
             buy_condition.append(df[f"mfi_{i}"].between(0, 7 + i))
             buy_condition.append(df[f"cti_{i}"].between(-0.95, (-0.90 + i / 100)))
             buy_condition.append(df[f"cti_{i-1}"].ge(df[f"cti_{i}"]))
+            buy_condition.append(df['lowertail'].ge(1.002))
 
             buy = reduce(lambda x, y: x & y, buy_condition)
             df.loc[buy, 'buy_tag'] = f"buy_{i}"
@@ -126,11 +114,10 @@ class ViN(IStrategy):
                 self.custom_buy_info[buy_date]['buy_tags'] += row['buy_tag']
                 self.custom_buy_info[buy_date]['buy_signals'] += 1
 
-        if self.config['runmode'].value not in ('live', 'dry_run'):
-            if self.write_to_csv:
-                df['pair'] = metadata['pair']
-                with open(self.df_csv, 'a') as f:
-                    df.to_csv(f, sep='\t', header=f.tell()==0, index=False)
+        if self.config['runmode'].value not in ('live', 'dry_run') and self.write_to_csv:
+            df['pair'] = metadata['pair']
+            with open(self.df_csv, 'a') as f:
+                df.to_csv(f, sep='\t', header=f.tell()==0, index=False)
 
         return df
 
@@ -139,43 +126,66 @@ class ViN(IStrategy):
             return df
 
         df.loc[:, 'sell_tag'] = ''
-        sell_up_time_periods = range(6, 16)
-        for i in sell_up_time_periods:
-            sell_condition = []
-            if i == min(sell_up_time_periods):
-                sell_condition.append(df['streak_max'].between(3, i))
-            elif i == max(sell_up_time_periods):
-                sell_condition.append(df['streak_max'].ge(i))
-            else:
-                sell_condition.append(df['streak_max'].eq(i))
-            sell_condition.append(df[f"rsi_{i}"].between(92 - i, 100))
-            sell_condition.append(df[f"mfi_{i}"].between(97 - i, 100))
-            sell_condition.append(df[f"cti_{i}"].ge((i - 6) / 10))
-            sell_condition.append(df[f"cti_{i-1}"].le(df[f"cti_{i}"]))
-
-            sell = reduce(lambda x, y: x & y, sell_condition)
-            df.loc[sell, 'sell_tag'] = f"sell_up_{i}"
-
         try:
             candles_between = df.index[-1] - df.loc[df['buy_tag'] != ''].index[-1]
         except:
             candles_between = self.startup_candle_count
-        sell_down_time_periods = range(4, min(candles_between, 16))
-        for i in sell_down_time_periods:
-            sell_condition = []
-            if i == min(sell_down_time_periods):
-                sell_condition.append(df['streak_min'].between(-i, -3))
-            elif i == max(sell_down_time_periods):
-                sell_condition.append(df['streak_min'].le(-i))
+        time_periods = range(5, min(candles_between, 16))
+        for i in time_periods:
+            sell_conditions = []
+            sell_conditions.append(df['candle_count'].ge(self.startup_candle_count))
+            sell_conditions.append(df['volume'].ge(self.min_vol_candle))
+            sell_conditions.append(df['volume_6'].ge(self.min_vol_6_candles))
+            if i == min(time_periods):
+                sell_conditions.append(df['streak_max'].between(3, i))
+            elif i == max(time_periods):
+                sell_conditions.append(df['streak_max'].ge(i))
             else:
-                sell_condition.append(df['streak_min'].eq(-i))
-            sell_condition.append(df[f"rsi_{i}"].between(43 - i, 100))
-            sell_condition.append(df[f"mfi_{i}"].between(35 - i, 100))
-            sell_condition.append(df[f"cti_{i}"].le((-i + 6) / 10))
-            sell_condition.append(df[f"cti_{i-1}"].gt(df[f"cti_{i}"]))
+                sell_conditions.append(df['streak_max'].eq(i))
+            # sell_conditions.append(df[f"_{i}"].between(74 - i, 100)) # 92
+            sell_conditions.append(df[f"mfi_{i}"].between(96 - i, 100)) # 97
+            sell_conditions.append(df[f"cti_{i}"].ge((i - 6) / 10))
+            sell_conditions.append(df[f"cti_{i-1}"].le(df[f"cti_{i}"]))
 
-            sell = reduce(lambda x, y: x & y, sell_condition)
+            sell = reduce(lambda x, y: x & y, sell_conditions)
+            df.loc[sell, 'sell_tag'] = f"sell_up_{i}"
+
+        time_periods = range(5, min(candles_between, 16))
+        for i in time_periods:
+            sell_conditions = []
+            sell_conditions.append(df['candle_count'].ge(self.startup_candle_count))
+            sell_conditions.append(df['volume'].ge(self.min_vol_candle))
+            # sell_conditions.append(df['volume_6'].ge(self.min_vol_6_candles))
+            if i == min(time_periods):
+                sell_conditions.append(df['streak_min'].between(-i, -3))
+            elif i == max(time_periods):
+                sell_conditions.append(df['streak_min'].le(-i))
+            else:
+                sell_conditions.append(df['streak_min'].eq(-i))
+            sell_conditions.append(df[f"mfi_{i}"].between(35 - i, 100))
+            sell_conditions.append(df[f"cti_{i}"].le((-i + 6) / 10))
+            sell_conditions.append(df[f"cti_{i-1}"].gt(df[f"cti_{i}"]))
+            sell_conditions.append(df[f"cti_2"].lt(0.01))
+
+            sell = reduce(lambda x, y: x & y, sell_conditions)
             df.loc[sell, 'sell_tag'] = f"sell_down_{i}"
+
+        time_periods = range(3, min(candles_between, 16))
+        for i in time_periods:
+            sell_conditions = []
+            sell_conditions.append(df['candle_count'].ge(self.startup_candle_count))
+            if i == min(time_periods):
+                sell_conditions.append(df['streak_min'].between(-i, -2))
+            elif i == max(time_periods):
+                sell_conditions.append(df['streak_min'].le(-i))
+            else:
+                sell_conditions.append(df['streak_min'].eq(-i))
+            sell_conditions.append(df[f"mfi_{i}"].between(35 - i, 100))
+            sell_conditions.append(df[f"cti_{i}"].le((-i + 6) / 10))
+            sell_conditions.append(df[f"cti_{i-1}"].gt(df[f"cti_{i}"]))
+
+            sell = reduce(lambda x, y: x & y, sell_conditions)
+            df.loc[sell, 'sell_tag'] += f"stop_{i}"
 
         df.loc[:, 'sell'] = False
         return df
@@ -201,15 +211,17 @@ class ViN(IStrategy):
         current_profit = (current_rate - trade.open_rate) / trade.open_rate
         # max_profit = (max_rate - trade.open_rate) / trade.open_rate
 
-        if current_profit > 0.015 or current_profit < -0.015:
-            sell_tag = candle_1['sell_tag']
-            if sell_tag != '':
-                if candle_1['buy']:
-                    log.info(f"custom sell: sell cancelled with sell_tag {candle_1['sell_tag']} and buy_tag {candle_1['buy_tag']}")
-                    return None
-                else:
-                    candle_1['sell'] = True
-                    return f"{sell_tag} ({buy_tag})"
+        sell_tag = candle_1['sell_tag']
+        if sell_tag != '':
+            if candle_1['buy']:
+                log.info(f"custom_sell: sell for pair {pair} with sell_tag {sell_tag} and buy_tag {buy_tag} on candle {candle_1['date']} is cancelled.")
+                return None
+            elif (
+                (current_profit < -0.015 and 'stop' in sell_tag)
+                or (current_profit > 0.015 and 'sell' in sell_tag)
+            ):
+                candle_1['sell'] = True
+                return f"{sell_tag} ({buy_tag})"
 
         return None
 
@@ -233,15 +245,17 @@ class ViN(IStrategy):
         buy_candle_date = candle_1['date']
 
         if buy_candle_date in self.custom_buy_info.keys():
+            pairs = len(self.dp.current_whitelist())
+            max_concurrent_buy_signals = np.floor(pairs * 0.08)
             buy_tags = self.custom_buy_info[buy_candle_date][pair]
-            if  self.max_concurrent_buy_signals > 0:
+            if  max_concurrent_buy_signals > 0:
                 buy_info = self.custom_buy_info[buy_candle_date]
                 buy_signal_count = buy_info['buy_signals']
-                if buy_signal_count > self.max_concurrent_buy_signals:
-                    log.info(f"confirm_trade_entry: Buy for pair {pair} with buy tag {buy_tags} on candle {buy_candle_date} is cancelled. There are {buy_signal_count} concurrent buy signals (max = {self.max_concurrent_buy_signals}).")
+                if buy_signal_count > max_concurrent_buy_signals:
+                    log.info(f"confirm_trade_entry: Buy for pair {pair} with buy tag {buy_tags} on candle {buy_candle_date} is cancelled. There are {buy_signal_count} concurrent buy signals (max = {max_concurrent_buy_signals}).")
                     return False
 
-            if self.config['runmode'].value not in ('live', 'dry_run'):
+            if self.config['runmode'].value not in ('live', 'dry_run') and self.write_to_txt:
                 close_1_price = candle_1['close']
                 close_2_price = candle_2['close']
                 indicator = []
@@ -264,19 +278,18 @@ class ViN(IStrategy):
 
     def confirm_trade_exit(self, pair: str, trade: "Trade", order_type: str, amount: float,
                            rate: float, time_in_force: str, sell_reason: str, **kwargs) -> bool:
-        df, _ = self.dp.get_analyzed_dataframe(pair, self.timeframe)
-        df: DataFrame = df
-
-        if self.config['runmode'].value not in ('live', 'dry_run'):
+        if self.config['runmode'].value not in ('live', 'dry_run') and self.write_to_txt:
+            df, _ = self.dp.get_analyzed_dataframe(pair, self.timeframe)
+            df: DataFrame = df
             trade_open_date = timeframe_to_prev_date(self.timeframe, trade.open_date_utc)
             trade_close_date = timeframe_to_prev_date(self.timeframe, trade.close_date_utc)
-            buy_tag = trade.buy_tag if trade is not None else 'empty'
+            df_trade = df.loc[(df['date'].ge(trade_open_date)) & (df['date'].le(trade_close_date))]
 
-            df_trade = df[(df['date'].ge(trade_open_date)) & (df['date'].le(trade_close_date))]
             if df_trade.empty:
                 log.warning(f"confirm_trade_exit: Empty trade dataframe for pair {pair} on trade with open date {trade_open_date} and close date {trade_close_date}.")
                 return False
 
+            buy_tag = trade.buy_tag if trade is not None else 'empty'
             max_close_candle = df_trade.nlargest(1, columns=['close'])
             min_close_candle = df_trade.nsmallest(1, columns=['close'])
             min_close_date = min_close_candle['date'].to_numpy()[0]
@@ -284,7 +297,6 @@ class ViN(IStrategy):
             profit = (rate - trade.open_rate) / trade.open_rate
             max_profit = (trade.max_rate - trade.open_rate) / trade.open_rate
             max_loss = (trade.min_rate - trade.open_rate) / trade.open_rate
-
             with open(self.f_trades, 'a') as f:
                 print(f'{pair};{trade_open_date};{trade.open_rate:.10n};{trade_close_date};{rate:.10n};{buy_tag};{sell_reason.partition(" (")[0]};{profit:.10n};{max_profit:.10n};{max_loss:.10n};{trade.max_rate:.10n};{trade.min_rate:.10n};{max_close_date};{min_close_date};', file=f)
